+++ CoreAnimationFunHouse/CALayer+WiggleAnimationAdditions.m
@@ -31,8 +31,8 @@ static NSString * const kBTSWiggleTransformTranslationXAnimation = @"BTSWiggleTr
     [rotationAnimation setDuration:0.1];
     [rotationAnimation setAutoreverses:YES];
 
-    [rotationAnimation setFromValue:[NSNumber numberWithFloat:M_PI/100]];
-    [rotationAnimation setToValue:[NSNumber numberWithFloat:-M_PI/100]];
+    [rotationAnimation setFromValue:[NSNumber numberWithFloat:(float) (M_PI/100.0)]];
+    [rotationAnimation setToValue:[NSNumber numberWithFloat:(float) (-M_PI/100.0)]];
     
     // Create the translation animation along the X axis. This gives is a slight sliding effect, which looks nice.
     CABasicAnimation *translationXAnimation = [CABasicAnimation animationWithKeyPath:@"transform.translation.x"];
@@ -40,8 +40,8 @@ static NSString * const kBTSWiggleTransformTranslationXAnimation = @"BTSWiggleTr
     [translationXAnimation setDuration:0.2];
 
     [translationXAnimation setAutoreverses:YES];
-    [translationXAnimation setFromValue:[NSNumber numberWithFloat:[self bounds].origin.x + 2]];
-    [translationXAnimation setToValue:[NSNumber numberWithFloat:[self bounds].origin.x - 2]];
+    [translationXAnimation setFromValue:[NSNumber numberWithFloat:[self bounds].origin.x + 2.0]];
+    [translationXAnimation setToValue:[NSNumber numberWithFloat:[self bounds].origin.x - 2.0]];
 
     // add the animations using app-specific keys... we use these keys to "stop wiggling".
     [self addAnimation:rotationAnimation forKey:kBTSWiggleTransformAnimation];
